Move createStore to store
